[修复] 1. 修复nlog.config引用不存在的target导致日志丢失的问题 [修复] 2. 修复ExceptionReportModel与数据库NULL约束不一致的问题 [优化] 3. 将DbContext改为DbContextPool并添加重试策略提升性能 [优化] 4. 优化NewsGenerateService中的异步模式（Thread.Sleep改为Task.Delay，移除Task.Run） [优化] 5. 为API查询添加AsNoTracking提升查询性能 [重构] 6. 将uint主键改为long避免类型兼容问题 [修复] 7. 修正StatsController中的HTTP状态码使用（500改为404/400） [文档] 8. 清理appsettings.json中的无效Serilog配置 [优化] 9. 添加数据库索引提升查询性能 [功能] 10. 添加全局异常处理中间件

+++ TactiX_Server/Models/Stats/VersionControlModel.cs
@@ -10,7 +10,7 @@ namespace TactiX_Server.Models.Stats
         /// 主索引
         /// </summary>
         [Key]
-        public uint Id { get; set; }
+        public long Id { get; set; }
         /// <summary>
         /// 版本号
         /// </summary>
